[Refactor]: Clean duplicated reduce mean function (#4056)
