+++ mmdet/core/utils/__init__.py
@@ -1,4 +1,7 @@
-from .dist_utils import DistOptimizerHook, allreduce_grads
+from .dist_utils import DistOptimizerHook, allreduce_grads, reduce_mean
 from .misc import multi_apply, unmap
 
-__all__ = ['allreduce_grads', 'DistOptimizerHook', 'multi_apply', 'unmap']
+__all__ = [
+    'allreduce_grads', 'DistOptimizerHook', 'reduce_mean', 'multi_apply',
+    'unmap'
+]
