+++ mmdet/core/utils/dist_utils.py
@@ -58,3 +58,12 @@ class DistOptimizerHook(OptimizerHook):
         warnings.warn('"DistOptimizerHook" is deprecated, please switch to'
                       '"mmcv.runner.OptimizerHook".')
         super().__init__(*args, **kwargs)
+
+
+def reduce_mean(tensor):
+    """"Obtain the mean of tensor on different GPUs."""
+    if not (dist.is_available() and dist.is_initialized()):
+        return tensor
+    tensor = tensor.clone()
+    dist.all_reduce(tensor.div_(dist.get_world_size()), op=dist.ReduceOp.SUM)
+    return tensor
